feat: use text colour for line & selection highlights
BREAKING CHANGE:

Previously, the current line highlight and selection highlight used the accent (red) colour. While it seemed like a good idea in the beginning due to various consistency reasons, it turned out to be rather a bad choice - red is also used in the diff view and selecting anything in the diff view, or simply trying to find the line where the cursor is currently placed by colour was almost impossible as they blended completely with the diff.

Changing the highlights to a very light variant of the base text colour improves the overal experience quite significantly. 🎨