feat: use orange for all accents
BREAKING CHANGE:

Orange is the new red. 😍

The reddish accents were great and all but the accents frequently clashed with diffs, errors and other things which are expected to be red by their very nature. Using a different accent we are free of those clashes. Don't worry, though, the orange still looks awesome! 🎨