feat: apply theme colours to input borders
Some inputs and dropdowns did not have a border before this change and some inputs and dropdowns used a VS Code-provided generic gray which was especially noticeable on the light theme variant.

With this change we fix the gray borders to use theme colours but we also add borders to all inputs. It actually improves UX a bit by making the input boxes' bounds much more clear.