feat: overhaul CSS/Less language colouring 🎨
My goal was to bring the colours on par with HTML and other, regular programming languages. So, HTML classes are now the same colour as ie. JavaScript classes etc. It's a bit less colourful now, I guess, but hopefully it will help you with navigating the files by instantly recognising most of the associated colours.